Organization of file structures , Adding final keyword to classes
